+++ server/component/style/validate/tpl_validate.php
@@ -18,7 +18,7 @@
                     <label>Leave this field empty</label>
                     <input type="text" class="form-control" name="phone7h92jP" autocomplete="off">
                 </div>
-                <div class="form-group">
+                <div class="form-group <?php echo $this->get_css_name(); ?>">
                     <label><?php echo $this->name_label; ?></label>
                     <input type="text" class="form-control" autocomplete="username" name="name" placeholder="<?php echo $this->name_placeholder; ?>" value="<?php echo $name; ?>" required>
                     <small class="form-text text-muted"><?php echo $this->name_description; ?></small>
@@ -28,7 +28,7 @@
                     <input type="password" autocomplete="new-password" class="form-control mb-1" name="pw" placeholder="<?php echo $this->pw_placeholder; ?>" required>
                     <input type="password" autocomplete="new-password" class="form-control" name="pw_verify" placeholder="<?php echo $this->pw_confirm_label; ?>" required>
                 </div>
-                <div class="form-group">
+                <div class="form-group <?php echo $this->get_css_gender(); ?>">
                     <div>
                     <label><?php echo $this->gender_label; ?></label>
                     </div>